+++ src/main/kotlin/com/github/strindberg/emacsj/actions/movement/GotoLineAction.kt
@@ -1,0 +1,6 @@
+package com.github.strindberg.emacsj.actions.movement
+
+import com.github.strindberg.emacsj.movement.GotoLineHandler
+import com.intellij.openapi.editor.actions.TextComponentEditorAction
+
+class GotoLineAction : TextComponentEditorAction(GotoLineHandler())