+++ src/test/kotlin/com/github/strindberg/emacsj/movement/ToIndentationTest.kt
@@ -1,6 +1,5 @@
 package com.github.strindberg.emacsj.movement
 
-import com.github.strindberg.emacsj.word.FILE
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 
 class ToIndentationTest : BasePlatformTestCase() {